requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via -r requirements.in
smmap==5.0.0
    # via gitdb
